reworking requests
