+++ backend/src/commands/properties.js
@@ -5,5 +5,5 @@ dotenv.config();
 const conn = mysql.createConnection({
   host: process.env.AWS_RDS_HOST,
   user: process.env.AWS_RDS_USER,
-  password: process.env.AWS_RDS_PASSWORD
-})
+  password: process.env.AWS_RDS_PASSWORD,
+});
